+++ src/com/elfec/ssc/model/webservices/converters/GetAllLocationPointsConverter.java
@@ -24,7 +24,7 @@ public class GetAllLocationPointsConverter implements
 					payPoints.add(new LocationPoint(object.getString("institution_name"),object.getString("address"), 
 							object.getString("phone"), object.getString("start_attention"), 
 							object.getString("end_attention"), object.getDouble("latitude"), 
-							object.getDouble("longitude"),Short.parseShort(object.getString("type"))));
+							object.getDouble("longitude"),(short)Short.parseShort(object.getString("type"))));
 				}
 			} catch (JSONException e) {
 				e.printStackTrace();
